Attempt to fix override-channels and dependencies
Prior to this commit, if bonito is ran on a fresh machine with no .lock
file and override-channels = true, when trying to add a channel that
depends on Nixpkgs (e.g. Home Manager), an error will occur as bonito
will have already removed the channel.

This commit attempts to fix this quirk by resolving the locks of missing
channels first, then it goes ahead with removing old channels and adding
new ones in.